refactor: move desktop runtime into ui package
