+++ src/cogstash/ui/__init__.py
@@ -1,5 +1,1 @@
-"""UI package scaffolding for future extraction work."""
-
-from __future__ import annotations
-
-__all__: list[str] = []
+"""Desktop UI package for CogStash."""
