+++ src/cogstash/ui/__main__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from cogstash.ui.app import main
+
+if __name__ == "__main__":
+    main()
