+++ tests/ui/conftest.py
@@ -1,0 +1,5 @@
+"""UI pytest fixtures and defensive desktop runtime stubs."""
+
+from __future__ import annotations
+
+from ui._support import *  # noqa: F401,F403
